requirements.txt
colorama==0.4.6
discord.py==2.1.0
requests==2.26.0